chore: remove and reorganize packages and tests folders
