--- cypress/support/e2e.ts
@@ -1,2 +1,0 @@
-import "@cypress/code-coverage/support";
-import "./commands";
