+++ package.json
@@ -3,42 +3,50 @@
 	"private": true,
 	"version": "0.0.0",
 	"type": "module",
+	"packageManager": "pnpm@8.11.0",
+	"engines": {
+		"node": "20.x.x"
+	},
 	"scripts": {
+		"prepare": "husky install",
 		"dev": "vite",
 		"build": "vite build",
 		"preview": "vite preview",
-		"format": "prettier --check src/",
-		"lint": "eslint \"src/**/*.{js,ts,vue}\"",
-		"type-check": "vue-tsc --noEmit",
-		"cy:open": "cypress open",
-		"cy:run": "cypress run"
+		"format": "prettier --check \"{src,tests}/**/*.{js,ts,vue}\"",
+		"lint": "eslint \"{src,tests}/**/*.{js,ts,vue}\"",
+		"type:check": "vue-tsc --noEmit",
+		"test:e2e": "cypress run --e2e --quiet",
+		"test:e2e:ui": "cypress open --e2e",
+		"test:component": "cypress run --component --quiet",
+		"test:component:ui": "cypress open --component"
 	},
 	"dependencies": {
-		"daisyui": "^3.9.3",
+		"daisyui": "^4.4.19",
 		"pinia": "^2.1.7",
-		"vue": "^3.3.6",
+		"vue": "^3.3.10",
 		"vue-router": "^4.2.5"
 	},
 	"devDependencies": {
-		"@cypress/code-coverage": "^3.12.4",
 		"@pinia/testing": "^0.1.3",
-		"@typescript-eslint/eslint-plugin": "^6.8.0",
-		"@typescript-eslint/parser": "^6.8.0",
-		"@vitejs/plugin-vue": "^4.4.0",
+		"@typescript-eslint/eslint-plugin": "^6.13.2",
+		"@typescript-eslint/parser": "^6.13.2",
+		"@vitejs/plugin-vue": "^4.5.2",
 		"autoprefixer": "^10.4.16",
-		"cypress": "^13.3.2",
-		"eslint": "^8.51.0",
-		"eslint-config-prettier": "^9.0.0",
-		"eslint-plugin-vue": "^9.17.0",
-		"nyc": "^15.1.0",
-		"postcss": "^8.4.31",
-		"prettier": "^3.0.3",
-		"prettier-plugin-tailwindcss": "^0.5.6",
-		"tailwindcss": "^3.3.3",
-		"typescript": "^5.2.2",
-		"vite": "^4.5.0",
-		"vite-plugin-istanbul": "^5.0.0",
+		"cypress": "^13.6.1",
+		"eslint": "^8.55.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-chai-friendly": "^0.7.2",
+		"eslint-plugin-cypress": "^2.15.1",
+		"eslint-plugin-vue": "^9.19.2",
+		"husky": "^8.0.3",
+		"lint-staged": "^15.2.0",
+		"postcss": "^8.4.32",
+		"prettier": "^3.1.0",
+		"prettier-plugin-tailwindcss": "^0.5.9",
+		"tailwindcss": "^3.3.6",
+		"typescript": "^5.3.3",
+		"vite": "^5.0.6",
 		"vue-eslint-parser": "^9.3.2",
-		"vue-tsc": "^1.8.19"
+		"vue-tsc": "^1.8.25"
 	}
 }
