+++ tests/support/component.ts
@@ -1,4 +1,3 @@
-import "@cypress/code-coverage/support";
 import "./commands";
 import "@src/style.css";
 
@@ -7,14 +6,6 @@ import { mount } from "cypress/vue";
 import { routes } from "@src/routes";
 import { createRouter, createWebHashHistory } from "vue-router";
 
-declare global {
-  namespace Cypress {
-    interface Chainable {
-      mount: typeof mount;
-    }
-  }
-}
-
 Cypress.Commands.add("mount", (component, options = {}) => {
   options.global = options.global || {};
   options.global.plugins = options.global.plugins || [];
@@ -22,7 +13,7 @@ Cypress.Commands.add("mount", (component, options = {}) => {
     createRouter({
       routes,
       history: createWebHashHistory(),
-    })
+    }),
   );
 
   return mount(component, {
