+++ tests/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../tsconfig.json",
+    "compilerOptions": {
+        "types": ["cypress", "./cypress.d.ts"]
+    },
+    "include": ["**/**.ts"]
+}
