+++ tsconfig.json
@@ -5,6 +5,9 @@
       "@src/*": [
         "./src/*"
       ],
+      "@tests/*": [
+        "./tests/*"
+      ]
     },
     "target": "ES2020",
     "useDefineForClassFields": true,
@@ -29,7 +32,6 @@
     "noFallthroughCasesInSwitch": true,
   },
   "include": [
-    "cypress",
     "src/**/*.ts",
     "src/**/*.d.ts",
     "src/**/*.tsx",
